apps/platform/server/pyproject.toml
[project]
name = "envel_platform"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "itsdangerous",
    "bcrypt",
    "python-json-logger",
    "sqlcipher3-wheels",
    "langgraph>=1.1.6",
    "langchain-openrouter>=0.2.1",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "httpx>=0.27",
]

[project.scripts]
envel-platform = "envel_platform.server:main"

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
